+++ apps/dialtone-documentation/docs/.vuepress/baseComponents/SvgLoader.vue
@@ -1,6 +1,13 @@
 <template>
-  <template v-for="svg in svgs" :key="svg">
-    <component :is="svg" v-if="svg" />
+  <template v-if="illustration">
+    <template v-for="i in illustrationSVGs" :key="i">
+      <component :is="i" v-if="i" :class="class" />
+    </template>
+  </template>
+  <template v-else>
+    <template v-for="svg in svgs" :key="svg">
+      <component :is="svg" v-if="svg" :class="class" />
+    </template>
   </template>
 </template>
 
@@ -16,7 +23,29 @@ const props = defineProps({
     type: String,
     required: true,
   },
+  /**
+   * Class attribute to be added to the component.
+   */
+  class: {
+    type: [String, Array, Object],
+    default: '',
+  },
+
+  /**
+   * Whether the svg is an illustration.
+   */
+  illustration: {
+    type: Boolean,
+    default: false,
+  },
 });
+
+const illustrationSVGs = [
+  defineAsyncComponent({
+    loader: () => import(`../../../node_modules/@dialpad/dialtone-icons/dist/svg/illustrations/${props.name}.svg`),
+    onError: () => {},
+  }),
+];
 
 const svgs = [
   defineAsyncComponent({
@@ -29,10 +58,6 @@ const svgs = [
   }),
   defineAsyncComponent({
     loader: () => import(`../public/assets/images/favicons/${props.name}.svg?component`),
-    onError: () => {},
-  }),
-  defineAsyncComponent({
-    loader: () => import(`../../../node_modules/@dialpad/dialtone-css/lib/dist/svg/spot/${props.name}.svg?component`),
     onError: () => {},
   }),
 ];