+++ packages/dialtone-vue2/components/rich_text_editor/rich_text_editor.vue
@@ -27,9 +27,9 @@ import Underline from '@tiptap/extension-underline';
 import Text from '@tiptap/extension-text';
 import TextAlign from '@tiptap/extension-text-align';
 import Emoji from './extensions/emoji';
-import Link from './extensions/link';
-import { MentionPlugin, mentionRegex } from './extensions/mentions/mention';
-import { ChannelPlugin, channelRegex } from './extensions/channels/channel';
+import CustomLink from './extensions/custom_link';
+import { MentionPlugin } from './extensions/mentions/mention';
+import { ChannelPlugin } from './extensions/channels/channel';
 import { SlashCommandPlugin } from './extensions/slash_command/slash_command';
 import {
   RICH_TEXT_EDITOR_OUTPUT_FORMATS,
@@ -40,8 +40,6 @@ import {
 import mentionSuggestion from './extensions/mentions/suggestion';
 import channelSuggestion from './extensions/channels/suggestion';
 import slashCommandSuggestion from './extensions/slash_command/suggestion';
-import emojiRegex from 'emoji-regex';
-import { codeToEmojiData, emojiShortCodeRegex } from '@/common/emoji';
 
 export default {
   name: 'DtRichTextEditor',
@@ -124,7 +122,7 @@ export default {
      */
     outputFormat: {
       type: String,
-      default: 'text',
+      default: 'html',
       validator (outputFormat) {
         return RICH_TEXT_EDITOR_OUTPUT_FORMATS.includes(outputFormat);
       },
@@ -139,9 +137,27 @@ export default {
     },
 
     /**
-     * Enables the Link extension and optionally passes configurations to it
+     * Enables the TipTap Link extension and optionally passes configurations to it
+     *
+     * It is not recommended to use this and the custom link extension at the same time.
      */
     link: {
+      type: [Boolean, Object],
+      default: false,
+    },
+
+    /**
+     * Enables the Custom Link extension and optionally passes configurations to it
+     *
+     * It is not recommended to use this and the built in TipTap link extension at the same time.
+     *
+     * The custom link does some additional things on top of the built in TipTap link
+     * extension such as styling phone numbers and IP adresses as links, and allows you
+     * to linkify text without having to type a space after the link. Currently it is missing some
+     * functionality such as editing links and will likely require more work to be fully usable,
+     * so it is recommended to use the built in TipTap link for now.
+     */
+    customLink: {
       type: [Boolean, Object],
       default: false,
     },
@@ -243,6 +259,14 @@ export default {
       type: Boolean,
       default: true,
     },
+
+    /**
+     * Whether the input allows codeblock to be introduced in the text.
+     */
+    allowCodeblock: {
+      type: Boolean,
+      default: true,
+    },
   },
 
   emits: [
@@ -297,7 +321,16 @@ export default {
       // These are the default extensions needed just for plain text.
       const extensions = [Document, Paragraph, Text];
       if (this.link) {
-        extensions.push(this.getExtension(Link, this.link));
+        extensions.push(TipTapLink.extend({ inclusive: false }).configure({
+          HTMLAttributes: {
+            class: 'd-link d-wb-break-all',
+          },
+          autolink: true,
+          protocols: RICH_TEXT_EDITOR_SUPPORTED_LINK_PROTOCOLS,
+        }));
+      }
+      if (this.customLink) {
+        extensions.push(this.getExtension(CustomLink, this.customLink));
       }
       if (this.allowBlockquote) {
         extensions.push(Blockquote);
@@ -308,9 +341,7 @@ export default {
       if (this.allowBulletList) {
         extensions.push(BulletList);
         extensions.push(ListItem);
-        extensions.push(OrderedList.configure({
-          itemTypeName: 'listItem',
-        }));
+        extensions.push(OrderedList);
       }
       if (this.allowItalic) {
         extensions.push(Italic);
@@ -323,13 +354,16 @@ export default {
       }
 
       // Enable placeholderText
-      extensions.push(
-        Placeholder.configure({ placeholder: this.placeholder }),
-      );
+      if (this.placeholder) {
+        extensions.push(
+          Placeholder.configure({ placeholder: this.placeholder }),
+        );
+      }
 
       // make sure that this is defined before any other extensions
       // where Enter and Shift+Enter should have its own interaction. otherwise it will be ignored
       if (!this.allowLineBreaks) {
+        const self = this;
         extensions.push(
           HardBreak.extend({
             addKeyboardShortcuts () {
@@ -337,6 +371,7 @@ export default {
                 Enter: () => true,
                 'Shift-Enter': () => this.editor.commands.first(({ commands }) => [
                   () => commands.newlineInCode(),
+                  () => self.allowBulletList && commands.splitListItem('listItem'),
                   () => commands.createParagraphNear(),
                   () => commands.liftEmptyBlock(),
                   () => commands.splitBlock(),
@@ -346,11 +381,6 @@ export default {
           }),
         );
       }
-
-      extensions.push(TipTapLink.extend({ inclusive: false }).configure({
-        autolink: true,
-        protocols: RICH_TEXT_EDITOR_SUPPORTED_LINK_PROTOCOLS,
-      }));
 
       if (this.mentionSuggestion) {
         // Add both the suggestion plugin as well as means for user to add suggestion items to the plugin
@@ -379,11 +409,13 @@ export default {
         defaultAlignment: 'left',
       }));
 
-      extensions.push(CodeBlock.configure({
-        HTMLAttributes: {
-          class: 'dt-rich-text-editor--code-block',
-        },
-      }));
+      if (this.allowCodeblock) {
+        extensions.push(CodeBlock.configure({
+          HTMLAttributes: {
+            class: 'dt-rich-text-editor--code-block',
+          },
+        }));
+      }
 
       return extensions;
     },
@@ -438,9 +470,8 @@ export default {
         // through the parent, so don't do anything here.
         return;
       }
-
-      this.internalValue = newValue;
-      this.insertContent();
+      // Otherwise replace the content (resets the cursor position).
+      this.editor.commands.setContent(newValue, false);
     },
   },
 
@@ -461,6 +492,7 @@ export default {
       // For all available options, see https://tiptap.dev/api/editor#settings
       this.editor = new Editor({
         autofocus: this.autoFocus,
+        content: this.value,
         editable: this.editable,
         extensions: this.extensions,
         editorProps: {
@@ -470,73 +502,7 @@ export default {
           },
         },
       });
-      this.insertContent();
       this.addEditorListeners();
-    },
-
-    /**
-     * This function is necessary as tiptap doesn't render the content passed
-     * directly through `editor.commands.setContent` the content passed down to it
-     * should be already parsed. So We're parsing the elements into it's corresponding
-     * HTML version before setting it.
-     */
-    insertContent () {
-      this.parseMentions();
-      this.parseChannels();
-      this.parseEmojis();
-      this.editor.commands.setContent(this.internalValue, true);
-    },
-
-    parseEmojis () {
-      const matches = new Set(
-        [...this.value.matchAll(emojiRegex()), ...this.value.matchAll(emojiShortCodeRegex)]
-          .map(match => match[0].trim()),
-      );
-
-      if (!matches) return;
-
-      matches
-        .forEach(match => {
-          const emoji = codeToEmojiData(match);
-          if (!emoji) return;
-          this.internalValue = this.internalValue.replace(new RegExp(`${match}`, 'g'), `<emoji-component code="${emoji.shortname}"></emoji-component>`);
-        });
-    },
-
-    parseChannels () {
-      if (!this.channelSuggestion) return;
-
-      const suggestions = this.channelSuggestion.items({ query: '' });
-      const matches = [...this.value.matchAll(channelRegex)]
-        .filter(match => suggestions.some(({ id }) => id === match[1]));
-
-      if (!matches) return;
-
-      matches.forEach(match => {
-        const channel = suggestions.find(({ id }) => id === match[1]);
-        this.internalValue = this.internalValue.replace(
-          `#${match[1]}`,
-          /** The space at the beginning is important as tiptap removes that while rendering.
-           *  So if multiple mentions, channels or emojis are next to each other it will fail
-           */
-          ` <channel-component name="${channel.name}" id="${channel.id}"></channel-component>`,
-        );
-      });
-    },
-
-    parseMentions () {
-      if (!this.mentionSuggestion) return;
-
-      const suggestions = this.mentionSuggestion.items({ query: '' });
-      const matches = [...this.value.matchAll(mentionRegex)]
-        .filter(match => suggestions.some(({ id }) => id === match[1]));
-
-      if (!matches) return;
-
-      matches.forEach(match => {
-        const mention = suggestions.find(({ id }) => id === match[1]);
-        this.internalValue = this.internalValue.replace(`@${match[1]}`, ` <mention-component name="${mention.name}" id="${mention.id}"></mention-component>`);
-      });
     },
 
     destroyEditor () {
@@ -597,34 +563,40 @@ export default {
 </script>
 
 <style lang="less">
-  .ProseMirror p.is-editor-empty:first-child::before {
-    content: attr(data-placeholder);
-    float: left;
-    color: var(--dt-color-foreground-placeholder);
-    pointer-events: none;
-    height: 0;
-  }
-
-  .ProseMirror ul > li {
-    list-style-type: disc;
-  }
-
-  .ProseMirror ol > li {
-    list-style-type: decimal;
-  }
-
-  .ProseMirror blockquote {
-    padding-left: var(--dt-space-400);
-    border-left: var(--dt-size-border-300) solid var(--dt-color-foreground-muted-inverted);
-    margin-left: 0;
-  }
-
-  .dt-rich-text-editor--code-block {
-    background: var(--dt-color-surface-secondary);
-    padding: var(--dt-space-400);
-  }
-
   .dt-rich-text-editor {
-    overflow: hidden;
+    &--code-block {
+      background: var(--dt-color-surface-secondary);
+      padding: var(--dt-space-400);
+    }
+
+    > .ProseMirror {
+      box-shadow: none;
+
+      p.is-editor-empty:first-child::before {
+        content: attr(data-placeholder);
+        float: left;
+        color: var(--dt-color-foreground-placeholder);
+        pointer-events: none;
+        height: 0;
+      }
+
+      ul, ol {
+        padding-left: var(--dt-space-525);
+      }
+
+      ul > li {
+        list-style-type: disc;
+      }
+
+      ol > li {
+        list-style-type: decimal;
+      }
+
+      blockquote {
+        padding-left: var(--dt-space-400);
+        border-left: var(--dt-size-border-300) solid var(--dt-color-foreground-muted-inverted);
+        margin-left: 0;
+      }
+    }
   }
 </style>