+++ packages/dialtone-vue3/components/tooltip/modifiers.js
@@ -1,7 +1,0 @@
-export const getArrowDetected = fn => ({
-  name: 'arrowDetected',
-  enabled: true,
-  phase: 'main',
-  fn,
-  requiresIfExists: ['offset'],
-});