+++ packages/dialtone-vue3/recipes/leftbar/general_row/general_row.vue
@@ -401,10 +401,13 @@ export default {
     shouldApplyCustomStyleForCountBadge () {
       return this.hasUnreadCount && this.hasUnreadMentionCount;
     },
+<<<<<<< HEAD
 
     shouldApplyCustomStyleForMentionOnly () {
       return !this.hasUnreadCount && this.hasUnreadMentionCount;
     },
+=======
+>>>>>>> staging
   },
 
   watch: {